feat: separate localStorage logic
- 로컬 스토리지와 관련된 로직을 별도 파일로 분리합니다.
