+++ src/Logger.ts
@@ -1,5 +1,5 @@
-import { postLog } from './apis/postLog';
-import { LogPayloadParams, LogType, ServiceNameType } from './types/LogType';
+import { setLocalStorage } from './setLocalStorage';
+import { LogPayloadParams, ServiceNameType } from './types/LogType';
 
 const createHashedId = (userId: string) => {
   // Todo: create hashedId
@@ -10,28 +10,6 @@ const createTimestamp = () => {
   const offset = new Date().getTimezoneOffset() * 60 * 1000;
   const now = new Date(Date.now() - offset);
   return now.toISOString();
-};
-
-const setLocalStorageClear = () => {
-  const list: any[] = [];
-  localStorage.setItem('yls-web', JSON.stringify(list));
-};
-
-const setLocalStorage = async (logger: LogType) => {
-  if (window.localStorage.getItem('yls-web') == undefined) {
-    const list: any[] = [];
-    list.push(logger);
-    localStorage.setItem('yls-web', JSON.stringify(list));
-  } else {
-    const remainList: any[] = JSON.parse(localStorage.getItem('yls-web') as string) || [];
-    if (remainList.length < 10) {
-      const updateList = [...remainList, logger];
-      localStorage.setItem('yls-web', JSON.stringify(updateList));
-    } else {
-      setLocalStorageClear();
-      const res = await postLog();
-    }
-  }
 };
 
 const initialLog = (
